+++ robocorp-code/vscode-client/src/extension.ts
@@ -149,6 +149,7 @@ import {
     ROBOCORP_DEBUG_ACTION_FROM_ACTION_PACKAGE,
     ROBOCORP_CREATE_ACTION_PACKAGE,
     ROBOCORP_CREATE_TASK_OR_ACTION_PACKAGE,
+    ROBOCORP_DOWNLOAD_ACTION_SERVER,
 } from "./robocorpCommands";
 import { installPythonInterpreterCheck } from "./pythonExtIntegration";
 import { refreshCloudTreeView } from "./viewsRobocorp";
@@ -168,7 +169,7 @@ import { RobotOutputViewProvider } from "./output/outView";
 import { setupDebugSessionOutViewIntegration } from "./output/outViewRunIntegration";
 import { showInspectorUI } from "./inspector/inspectorView";
 import { IAppRoutes } from "./inspector/protocols";
-import { startActionServer } from "./actionServer";
+import { downloadLatestActionServer, startActionServer } from "./actionServer";
 import { askAndRunRobocorpActionFromActionPackage, createActionPackage } from "./robo/actionPackage";
 import { showSelectOneStrQuickPick } from "./ask";
 
@@ -389,6 +390,17 @@ function registerRobocorpCodeCommands(C: CommandRegistry, context: ExtensionCont
             } else if (packageType === ACTION_PACKAGE) {
                 createActionPackage();
             }
+        }
+    });
+    C.register(ROBOCORP_DOWNLOAD_ACTION_SERVER, async () => {
+        try {
+            const location = await downloadLatestActionServer();
+            window.showInformationMessage(`The latest action server was downloaded to: ${location}`);
+        } catch (error) {
+            logError("Error downloading latest action server", error, "ERR_DOWNLOAD_ACTION_SERVER");
+            window.showErrorMessage(
+                "There was an error downloading the action server. See `OUTPUT > Robocorp Code` for more information."
+            );
         }
     });
     C.register(ROBOCORP_UPLOAD_ROBOT_TO_CLOUD, () => uploadRobot());