+++ robocorp-code/vscode-client/src/robocorpCommands.ts
@@ -103,5 +103,6 @@ export const ROBOCORP_HAS_PRE_RUN_SCRIPTS_INTERNAL = "robocorp.hasPreRunScripts.
 export const ROBOCORP_RUN_PRE_RUN_SCRIPTS_INTERNAL = "robocorp.runPreRunScripts.internal";  // Run Pre Run Scripts
 export const ROBOCORP_GET_PY_PI_BASE_URLS_INTERNAL = "robocorp.getPyPiBaseUrls.internal";  // Get PyPi base urls
 export const ROBOCORP_START_ACTION_SERVER = "robocorp.startActionServer";  // Start Action Server
+export const ROBOCORP_DOWNLOAD_ACTION_SERVER = "robocorp.downloadActionServer";  // Download Action Server
 export const ROBOCORP_START_ACTION_SERVER_INTERNAL = "robocorp.startActionServer.internal";  // Start Action Server (internal)
 export const ROBOCORP_LIST_ACTIONS_INTERNAL = "robocorp.listActions.internal";  // Lists the actions available in an action package given a root dir (internal)