+++ robocorp-code/docs/changelog.md
@@ -1,3 +1,8 @@
+Unreleased
+-----------------------------
+
+- New action: `Robocorp Code: Download Action Server` (downloads the latest version of the Action Server).
+
 New in 1.18.0 (2024-02-29)
 -----------------------------